refactor: Enable reuse of a metric test macro
